Add crawlers and utils module
Crawlers will contain both the reddit
(already pressent) and the twitter
crawler classes.

Utils is an auxiliar file that I've created
to add functions that could be usefull in both
classes, such as get_random_string. But maybe
those functions could be defined on the fly
in the crawlers module.

I've also added a test script to try the code.