Implement Twitter multi file
Now if a tweet contains multiple photos, all of them
are sent one by one with the bot. Also, the image_name
and image_format parametres are no longer used, since
both of this things are infered from the url. This
ensures that we're saving the files with the correct
format.